pyproject.toml
[tool.poetry]
name = "wagtail-ninja"
version = "0.1.0"
description = "Wagtail CMS headless via ninja"
authors = ["Dan Brosnan <dpjbrosnan@octave.nz>"]
readme = "README.md"
packages = [{ include = "wagtail_ninja", from = "./src" }]

[tool.poetry.dependencies]
python = "^3.10"
Wagtail = ">=5.2"
django-ninja-extra = "0.21.6"
wagtail-headless-preview = ">=0.8"
pydantic = "^2"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "~7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
